Remove NA handling from pandas `read_csv`
This was a problem because the ISO code for Namibia is NA, which was getting
read in to the pd.DataFrame as an `NA`. That's obviously a problem (who thought
NA was a good idea for a 2 digit flag identifier anyway!?), so removed NA
handling with an argument. It's fine because the lookup csvs are curated anyway.