Fix Mode of Transport recoding logic to differentiate 2/3 digits
*Border* MoT is 3 digits, and *Inland* MoT is 2 digits. Didn't realise this
because they're both specified in the UKTradeInfo docs as 2 digits (not sure why
that leading zero got in there. Oh well). Rather than strip the leading zero, in
case of divergence in future, I've added in handling for the leading zero on the
recoding side which can then be simply changed down the line.