Add Checks for existing data to `monthly_update`
With a default threshold of 50000, you can simply query the database for number
of trade records in the time period specified in the arguments. **IF** there is
more than the threshold, then that months' data has been loaded into in the
past, so you shouldn't reload the trade data, otherwise you'll have to deal with
a bunch of deduplication headaches in the database that I'd rather not have to
handle every time I fat-finger an incorrect month for DB loads.